rtga: widen the fast-path row offset to size_t
The uncompressed 24/32-bit decode computes its destination row as

    uint32_t *dst = output + dst_row * tga_width;

with both operands int.  TGA dimensions are attacker-controlled and
run to 65535, so for a tall image dst_row * tga_width crosses INT_MAX
partway down - at row 32769 of a 65535-wide image - and the signed
multiply overflows.  That is undefined behaviour, and in practice the
product comes out negative, so the write lands before the buffer: a
wild out-of-bounds store.

The indexed path a little further down already computes the same
offset as (size_t)dst_row * (size_t)tga_width, with a comment noting
the 65535 x 65535 case.  The fast path was simply missed when that
fix went in.  Apply the identical widening.

Reaching it needs the output allocation - up to 16 GiB for the
largest dimensions - to succeed, which default Linux overcommit will
often permit, so this is reachable rather than theoretical.  The
signed overflow is UB regardless of whether the allocation backs.

Normal images are unaffected: the offset only differed once the int
product overflowed, well beyond any real asset.  A 64x48 decode is
byte-for-byte identical before and after.

+++ libretro-common/formats/tga/rtga.c
@@ -227,7 +227,10 @@ static uint32_t *rtga_tga_load(rtga_context *s,
       for (row = 0; row < tga_height; ++row)
       {
          int dst_row     = tga_inverted ? (tga_height - 1 - row) : row;
-         uint32_t *dst   = output + dst_row * tga_width;
+         /* size_t index so dst_row * tga_width does not overflow
+          * signed int for a legitimate 65535 x 65535 image - the
+          * same widening the indexed path below already does. */
+         uint32_t *dst   = output + (size_t)dst_row * (size_t)tga_width;
          int bytes_needed = tga_width * tga_comp;
          int col;
